Fix centering for baseFolder-not-found error
Previously, the error message would be placed in a div after the main
list table. Since the width was fixed, the error message would be
off-center. With this commit, the list table is only added to the doc if
the base folder was found, fixing the centering for the error message.